Project cleanup + integration stb

+++ Sources/labo.h
@@ -38,7 +38,7 @@ typedef struct {
 	Vector2 position;
 	uint8_t graph_group;
 	void* data;
-	uint8_t path_from;
+	uint64_t path_from;
 	uint8_t visited;
 } Node;
 typedef struct AdjMatrix AdjMatrix;
